requirements.txt
bidict==0.23.1
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
Flask==3.1.3
flask-cors==6.0.2
Flask-SocketIO==5.6.1
h11==0.16.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pyTelegramBotAPI==4.33.0
python-dotenv==1.2.2
python-engineio==4.13.1
python-socketio==5.16.1
requests==2.33.1
simple-websocket==1.1.0
telebot==0.0.5
urllib3==2.6.3
Werkzeug==3.1.8
wsproto==1.3.2
